requirements.txt
kivy>=2.1.0
Pillow>=9.0.0
